umbennen der funktionen

+++ GUI/motor_controlG.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import RPi.GPIO as GPIO
 import time
 
@@ -135,8 +137,8 @@ if __name__ == '__main__':
         while True:
             ##            delay = raw_input("Zeitverzoegerung (ms)?")
             steps = raw_input("Wie viele Schritte vorwaerts? ")
-            forward((int(delay) / 1000.0), int(steps))
+            forward((float(delay) / 1000.0), int(steps))
             steps = raw_input("Wie viele Schritte rueckwaerts? ")
-            backwards((int(delay) / 1000.0), int(steps))
+            backwards((float(delay) / 1000.0), int(steps))
     except KeyboardInterrupt:
         GPIO.cleanup()
